231110/requirements.txt
python-telegram-bot
openai
dotenv